Add support for plugin setup and finish actions
Enable plugins to have setup and finish functions that are run
before any tests are run and after all tests have run, respectively.
We added two decorators, plugin_setup and plugin_finish, to mark what
functions in a module are the setup and finish functions. The decorators
add that function as an attribute to the module object. Bandit.TestSet
then registers these functions as it load the plugins. Bandit.Manager
will run all setup functions for running the tests and it will run
all finish functions after it completes all tests.

Change-Id: I0122b3852e220d66d9f6d0426d4ed6312174dd17